pyproject.toml
[project]
name = "st-utils"
version = "0.2.0"
description = "Infrastructure and utilities for working with heterogeneous hazard-centric sensor data."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "lnetatmo>=4.2.0",
    "paho-mqtt>=2.1.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pytest>=8.3.4",
]

[tool.pytest.ini_options]
testpaths = [
    "tests"
]
filterwarnings = [
    "ignore:The 'User' class is no longer maintained by Netatmo:DeprecationWarning",
    "ignore:The 'DeviceList' class was renamed 'WeatherStationData':DeprecationWarning",
    "ignore:The 'WelcomeData' class was renamed 'HomeData':DeprecationWarning"
]
